Extract KeyButton and KeyButtonContextMenu from KeyGrid.py
